+++ tests/dev/script-outputs.R
@@ -3,14 +3,14 @@ sim <- loadSimulation("tests/data/S1.pkml")
 
 sim$solver$absTol <- 1e-13
 
-individualResults <- runSimulation(sim)
+individualResults <- runSimulations(sim)
 resultsPaths <- individualResults$allQuantityPaths
 
 tlf <- getOutputValues(individualResults)
 print(tlf)
 
 population <- loadPopulation("tests/data/pop_10.csv")
-populationResults <- runSimulation(sim, population)
+populationResults <- runSimulations(sim, population)
 resultsPaths <- populationResults$allQuantityPaths
 
 path <- resultsPaths[[1]]