+++ tests/testthat/test-simulation-results.R
@@ -1,7 +1,7 @@
 context("SimulationResults")
 
 sim <- loadTestSimulation("S1")
-simResults <- runSimulation(sim)
+simResults <- runSimulations(sim)
 
 resultsPath <- "Organism|PeripheralVenousBlood|Caffeine|Plasma (Peripheral Venous Blood)"