+++ tests/testthat/test-utility-pk-parameter.R
@@ -25,7 +25,7 @@ test_that("It calculates the pk parameters in the expected units", {
   myCmax <- addUserDefinedPKParameter(name = "MyCMax", standardPKParameter = StandardPKParameter$C_max, displayUnit = "mg/l")
   quantityPath <- "Organism|PeripheralVenousBlood|Caffeine|Plasma (Peripheral Venous Blood)"
   mw <- sim$molWeightFor(quantityPath)
-  results <- runSimulation(sim)
+  results <- runSimulations(sim)
   pkAnalyses <- calculatePKAnalyses(results)
   c_max_base_value <- pkAnalyses$pKParameterFor(quantityPath, c_max_base$name)$values[1]
   c_max_mg_l_value <- c_max_base_value * mw * 1E6